pipeline/custom_prediction/requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.0
google-cloud-aiplatform[prediction]==1.31.0